Only install java if not there on alinux/rhel

+++ core_deps.sh
@@ -11,20 +11,30 @@ then
   sudo yum install -y python27
   sudo yum install -y git
   sudo yum install -y vim
-  sudo yum install -y gcc openssl-devel readline-devel zlib-devel
+  sudo yum install -y gcc openssl-devel readline-devel zlib-devel gcc-c++
   git clone https://github.com/sstephenson/rbenv.git ~/.rbenv
   git clone https://github.com/sstephenson/ruby-build.git ~/.rbenv/plugins/ruby-build
 
   # Java
-  echo "Installing Java 1.7"
-  curl -v -j -k -L -H "Cookie: oraclelicense=accept-securebackup-cookie" http://download.oracle.com/otn-pub/java/jdk/7u79-b15/jdk-7u79-linux-x64.rpm > jdk-7u79-linux-x64.rpm
-  sudo rpm -Uvh jdk-7u79-linux-x64.rpm
-  rm jdk-7u79-linux-x64.rpm
+  if ! ls /usr/java/jdk1.7*
+  then
+    echo "Installing Java 1.7"
+    curl -v -j -k -L -H "Cookie: oraclelicense=accept-securebackup-cookie" http://download.oracle.com/otn-pub/java/jdk/7u79-b15/jdk-7u79-linux-x64.rpm > jdk-7u79-linux-x64.rpm
+    sudo rpm -Uvh jdk-7u79-linux-x64.rpm
+    rm jdk-7u79-linux-x64.rpm
+  else
+    echo "Java 1.7 already installed"
+  fi
 
-  echo "Installing Java 1.8"
-  curl -v -j -k -L -H "Cookie: oraclelicense=accept-securebackup-cookie" http://download.oracle.com/otn-pub/java/jdk/8u51-b16/jdk-8u51-linux-x64.rpm > jdk-8u51-linux-x64.rpm
-  sudo rpm -Uvh jdk-8u51-linux-x64.rpm
-  rm jdk-8u51-linux-x64.rpm
+  if ! ls /usr/java/jdk1.8*
+  then
+    echo "Installing Java 1.8"
+    curl -v -j -k -L -H "Cookie: oraclelicense=accept-securebackup-cookie" http://download.oracle.com/otn-pub/java/jdk/8u51-b16/jdk-8u51-linux-x64.rpm > jdk-8u51-linux-x64.rpm
+    sudo rpm -Uvh jdk-8u51-linux-x64.rpm
+    rm jdk-8u51-linux-x64.rpm
+  else
+    echo "Java 1.8 already installed"
+  fi
 fi
 
 if is_ubuntu
